// src/interfaces.ts
export interface IAccessToken {
  accessToken: string;
  tokenExpiry: Date;
}

export interface IMarketoLead {
  [key: string]: string | number | Date | boolean | null;
  id: number;
  email: string;
}

export interface ILeadValue {
  value: string | number | Date | boolean | null;
}

export interface IFieldList {
  [key: string]: any[];
  company: any[];
  billingCity: any[];
  billingCountry: any[];
  mainPhone: any[];
  id: any[];
  mktoName: any[];
  email: any[];
  phone: any[];
  address: any[];
  city: any[];
  country: any[];
  originalSourceType: any[];
  originalSourceInfo: any[];
  registrationSourceType: any[];
  registrationSourceInfo: any[];
  originalSearchEngine: any[];
  originalReferrer: any[];
  unsubscribedReason: any[];
  marketingSuspended: any[];
  mktoPersonNotes: any[];
  sfdcType: any[];
  anonymousIP: any[];
  inferredCompany: any[];
  inferredCountry: any[];
  inferredCity: any[];
  inferredStateRegion: any[];
  inferredPostalCode: any[];
  inferredPhoneAreaCode: any[];
  createdAt: any[];
  updatedAt: any[];
  leadSource: any[];
  leadStatus: any[];
  leadScore: any[];
  urgency: any[];
  priority: any[];
  mktoAcquisitionDate: any[];
  marketingSuspendedReason: any[];
  personTimeZone: any[];
  CurrencyIsoCode: any[];
  School__c: any[];
  School_Type__c: any[];
  Division__c: any[];
  Primary_Interest__c: any[];
  Email_Opt_Out__c: any[];
  Primary_Contact__c: any[];
  Unqualified_Reason__c: any[];
  Person_Type__c: any[];
  Region_Mapping__c: any[];
  Region__c: any[];
  Lead_Source_Category__c: any[];
  Integration_Source__c: any[];
  Intended_1st_year_of_study__c: any[];
  Parent_Email__c: any[];
  Parent_First_Name__c: any[];
  Parent_Last_Name__c: any[];
  Parent_Phone__c: any[];
  EventbriteSync__EventbriteId__c: any[];
  Birth_Date__c: any[];
  Eventbrite_Status__c: any[];
  Gender__c: any[];
  Sales_Funnel_Stage__c: any[];
  CurrencyIsoCode_account: any[];
  Currency__c: any[];
  CurrencyIsoCode_contact: any[];
  School_Country__c: any[];
  Account_Number__c: any[];
  New_Not_Contacted_Date_Time__c: any[];
  Contact_Number__c: any[];
  mkto_si__Add_to_Marketo_Campaign__c: any[];
  RecordTypeId: any[];
  Student_Email__c: any[];
  Student_First_Name__c: any[];
  Student_Last_Name__c: any[];
  End_Date__c: any[];
  Start_Date__c: any[];
  Behavioural_Score__c: any[];
  Status__c: any[];
  Sub_Status__c: any[];
  Programme__c: any[];
  Interests__c: any[];
  Division__c_account: any[];
  Primary_Language__c: any[];
  Secondary_Language__c: any[];
  RecordTypeId_account: any[];
  Payment_Method__c: any[];
  Payment_Status__c: any[];
  Year_of_Graduation__c: any[];
  Allocated__c: any[];
  Latest_NPS_Score__c: any[];
  Primary_Guardian__c: any[];
  Converted_from_MQ_to_SQ__c: any[];
  Converted_from_SQ_to_WD__c: any[];
  Student_Contact__c: any[];
  Application_Cycle__c_account: any[];
  Number_of_Touches__c: any[];
  Enquiry__c: any[];
  School_Year_Grade_Level__c: any[];
  Marketing_Qualified_Date__c: any[];
  Disqualified_Date__c: any[];
  OperatingRegion__c: any[];
  Sales_Accepted_Date__c: any[];
  privacyPolicy: any[];
  emailoptin: any[];
  LeadDataMigration__c: any[];
  Parent_Mobile__c: any[];
  Student_Mobile__c: any[];
  Student_Phone__c: any[];
  Payment_Contact__c: any[];
  First_MTA__c: any[];
  Last_MTA__c: any[];
  First_Lead_Source_Category__c: any[];
  First_Lead_Source__c: any[];
  Last_LeadSourceCategory__c: any[];
  Last_LeadSource__c: any[];
  First_Lead_Source_Detail__c: any[];
  Last_MKT_UTM_Campaign__c: any[];
  subscriptionCGA: any[];
  subscriptionCrimsonCore: any[];
  subscriptionMedView: any[];
  subscriptionAthletics: any[];
  subscriptionCrimsonAcademies: any[];
  subscriptionCrimsonRise: any[];
  subscriptionInternationalTours: any[];
  Marketed_by__c: any[];
  SalesFunnelStage__c: any[];
  Time_to_MarketingQualify__c: any[];
  Time_to_SalesAccepted__c: any[];
  Lead_Prioritisation__c: any[];
  MTA_Score__c: any[];
  Number_of_Touches_Formula__c: any[];
  CGA_Core_Split__c: any[];
  Total_MTA_Count_CGA__c: any[];
  Total_MTA_Count_CrimsonCore__c: any[];
  Last_Stage_before_DQ__c: any[];
  Status_Attempting_Contact__c: any[];
  Status_Contacted__c: any[];
  Status_Converted__c: any[];
  Status_Disqualified__c: any[];
  Status_IC_Booked__c: any[];
  Status_IC_Complete__c: any[];
  Status_New__c: any[];
  Create_CrimsonApp__c: any[];
  subscriptionPartnerships: any[];
  subscriptionTGCC: any[];
  subscriptionEssayComp: any[];
  LastNpsSurveySent__c: any[];
  Qualifies_For_NPS__c: any[];
  Academic_goals__c: any[];
  Admissions_goals__c: any[];
  Extracurricular_goals__c: any[];
  subscriptionEssayCompVisibilityRule: any[];
  subscriptionTGCCVisibilityRule: any[];
  Time_to_Attempting_Contact_Hours__c: any[];
  First_Touch_Detail__c: any[];
  Last_Modified_Date_Months__c: any[];
  DaScoopComposer__Groove_Convert__c: any[];
  DaScoopComposer__Groove_Log_a_Call__c: any[];
  DaScoopComposer__Groove_Create_Opportunity__c: any[];
  mobileSMSReady: any[];
  sfdcId: any[];
}